Renamed minid_client to minid
